backend/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
beautifulsoup4==4.14.3
billiard==4.2.4
cbor2==5.8.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fastapi==0.129.0
h11==0.16.0
httptools==0.7.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
kombu==5.6.2
msgpack==1.1.2
outcome==1.3.0.post0
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pyOpenSSL==25.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.2.0
requests==2.32.5
selenium==4.41.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
sqlparse==0.5.5
starlette==0.52.1
trio==0.33.0
trio-websocket==0.12.2
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.41.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==16.0
wsproto==1.3.2
zope.interface==8.2
